region: show 1.1 ISetIdCookie when using 1.1 labels
If a region contained a v1.1 namespace index, show -region was still
showing the 1.2 IsetCookie which was 0's. Show the 1.1 cookie in that
case.

Verified it matches the sysfs set_cookie for 1.1 and 1.2 indexes.

One gap to note is if there are no indexes, linux defaults to 1.1
while we default to 1.2. Should be fairly rare as we init namespace
index on create goal.

Signed-off-by: Juston Li <juston.li@intel.com>