create goal: Fix security state validation
ConvertSecurityBitmask() seems broken as it assumes the states are
exclusive. For instance, the security flag could be FROZEN + DISABLED
or FROZEN + ENABLED and both states are converted to FROZEN. Similar
problem occurs for NOT_SUPPORTED.

Modify IsConfiguringForCreateGoalAllowed() to just check the security
enabled bit since thats all we care about in this case.

Fixes #38

Signed-off-by: Juston Li <juston.li@intel.com>