Fix BSR function for non-functional dimms in UEFI
This change fixes fixes dump debug log as well.

There were several bugs. The for loop could exit without finding
a matching dimm to use. Additionally, it never used DIMM_FROM_NODE()
for pCurDimmInfoNode but instead casted to a DIMM_INFO *. Also the
for loop was not needed, GetDimmByPid() suffices.

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>