Enable explicit "-smbus" CLI override for all DCPMMs
The "-smbus" flag is intended to force all communication to occur over
the smbus interface only. This change enables this feature and required
a large number of code changes as the smbus interface was previously
assumed to be inaccessible in OS.

As part of this change, the Dimms global struct was updated to include
non-functional dimms in addition to functional dimms, as this
simplifies a lot of the logic.

Some bugs are still remaining (can't communicate with non-functional
DCPMMs without specifying "-smbus"), but will be fixed in a future
commit.

In addition, the logic for show -sensor was updated to allow
communication with unmanageable and non-functional dimms. This was
useful during testing as sensor data is uncached and there's no more
reason to exclude dimms we can communicate with.

We are moving away from the terms initialized and uninitialized and
replacing them with functional and non-functional to make meaning more
clear. More clear definitions are coming in a future commit.

Some .gitignore settings were applied, resulting in internal files
getting deleted. More changes coming soon.

Changed name of CheckSupportedConfigDimms to
ExcludeUnsupportedConfigDimms in VerifyTargetDimms to make meaning
more clear.

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>