Set initial Value for pointer to DIMM_PCD_INFO structure in GetPCD function
One of the input parameters to GetPcd function is a double pointer to the
DIMM_PCD_INFO structure. The function allocates DIMM_PCD_INFO structures and
returns them through this parameter. This change sets the pointer to
DIMM_PCD_INFO to NULL before the main function logic. This helps the cleanup
logic to take appropriate action if it is NULL before returning from function.

This fix is needed because when the GetPCD function was fuzz tested a crash
was observed where the cleanup logic was trying to free the pointer to
DIMM_PCD_INFO pointing to an invalid location because it was not set to NULL.

Signed-off-by: James Sushanth Anandraj <james.sushanth.anandraj@intel.com>