Handle uninitialized LSA condition in Get Regions when displaying regions HII page
Get regions function is called when displaying regions HII page. This function calls
another function to refresh namespaces. During this refresh if the LSA was not initialized
we were throwing error and exiting from get regions. This change handles this condition
more gracefully. This prevents the condition reported in the bug where regions HII
page was not displaying when dims are configured as App Direct and LSA is deleted.

Signed-off-by: James Sushanth Anandraj <james.sushanth.anandraj@intel.com>