Implement automatic passthrough interface selection
This change fixes several bugs relating to knowing what interface and
payload size to use to communicate with the DCPMM. Using the name of
"method" to refer to the possible combinations of the two,
DeterminePassThruMethod() will tell the caller what method to use based
on parameters passed on the CLI and several indicators located in the
pDimm struct.

The only use case so far to expose this logic outside of PassThru()
is to know whether the large payload interface is accessible so that
callers can determine if they need to split up their large payload
commands into many small payload "chunks". A new function,
IsLargePayloadAvailable() now exposes the necessary subset of this
logic.

From now on, there should be no need to check the indicators used in
DeterminePassThruMethod() except in very rare situations:
  * IS_DDRT_FLAG_ENABLED() in accessing the command access policy in
    CreateGoalConfig().
This change attempts to remove all references to the CLI/BSR parameters
located in our codebase and replace them with IsLargePayloadAvailable()
as appropriate.

Additionally, the functionality for accessing the BSR information and
setting the boot status bit mask was moved from NvmDimmConfig.c into
Dimm.c where it is more appropriate.

Also because of a previous change to the passthrough interface we no
longer have access to the mailbox status register, so we are removing
references to that field in the Show Register command and in our API.
This now allows us to run the Show Register command in OS as well,
though it continues to be debug build only.

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>