Update quick diagnostic to detect invalid boot status register
Quick diagnostic test reads the boot status register to perform
tests. If the value is invalid boot status register status variable
is set appropriately. This fix uses the status variable to determine
if boot status register value read is correct and makes sure that
diagnostic messages are updated if the value is incorrect.

Signed-off-by: James Sushanth Anandraj <james.sushanth.anandraj@intel.com>