Increase string size to match allowed input values in HII
In configure data policy page user can change average
power reporting time constant multiplier value and average power
reporting time constant by providing input as a string.
This fix increases the size of buffer that holds
these input strings so that it can hold the allowed input values.

In first case the allowed values are controlled by the macro
AVG_PWR_REPORTING_TIME_CONSTANT_MULT_MAX which has a value of 32.
So the input string buffer length is set to 3 to account
for 2 digits and additional end of string character.

In second case the allowed values are controlled by the macro
AVG_PWR_REPORTING_TIME_CONSTANT_MAX which has a value of 12000.
So the input string buffer length is set to 6 to account for 5
digits and additional end of string character.

Signed-off-by: James Sushanth Anandraj <james.sushanth.anandraj@intel.com>