Check Dimm count parameter against max dimms count for GetUninitializedDimms function
The GetUninitializedDimms function accepts a dimm count input parameter. During fuzz
testing of this API when large value for dimm count is passed a DXE Assert is observed.
This fix checks this input parameter against MAX_DIMMS macro which is set to 128 before
proceeding with the function to avoid DXE Assert.

Signed-off-by: James Sushanth Anandraj <james.sushanth.anandraj@intel.com>