Pass the force parameter when calling update firmware during firmware recovery
In the load command the update firmware function is called. One of the
parameters is the force option. The update firmware function checks if this
parameter is true before allowing firmware downgrade during recovery. Before
this change a hardcoded value of false was passed which prevented firmware
downgrade. By changing this value to a variable based on the arguments passed
by the user, a downgrade of firmware during recovery can now be achieved.

Signed-off-by: James Sushanth Anandraj <james.sushanth.anandraj@intel.com>