Update handling of Non-POR Dimms in show all dimms command and firmware update HII page
When Dimms are not in supported population configuration, the show all dimms command does
not display some properties such as manufacturer id. In Firmware update HII page the firmware
for these dimms are be updated.

This change fixes the above behaviors by removing the checks to see if dimms are in supported
population configuration in the show dimms command and firmware update HII page.

Signed-off-by: James Sushanth Anandraj <james.sushanth.anandraj@intel.com>