Improved error message when no usable DIMMs exist
When no usable DIMMs exist and a command (such as create-goal)
was run, the error message provided was simply 'DIMM not found'.
This message is confusing because DIMMs may have been able to be
found, they were just not usable. To improve this, a new message was
introduced for the case when DIMMs are found, but are not unusable.

Signed-off-by: Chris N. Johnson <chris.n.johnson@intel.com>