Handle no dcpmms when calling VerifyTargetDimms in get goal configs
Get goal configs is called when the region Hii page is opened. The get
goal configs then calls verifytargetdimms to get a list of dims that are
functional and not in population violation. When no dims are found it
returns EFI_NOT_FOUND with command status set to NVM_ERR_NO_USABLE_DIMMS.
This condition is not handled by get goal configs and returns error.
This causes region page to quit to main menu. This change enables the
get goal configs to recognize this condition and handle this situation.

Signed-off-by: James Sushanth Anandraj <james.sushanth.anandraj@intel.com>