Fix create goal issue when using -dimm to create goals on a subset of DIMMs
When the -dimm target is used in create goal to specify DCPMMs
on different sockets and the total number of DCPMMs specified
is equal to the number of DCPMMs on a socket, the call to the
IsConfigureWholeSocket() function inside of
VerifyCreatingSupportedRegionConfigs() will succeed incorrectly.
Instead of filtering out which specified DCPMMs are on which
sockets, removing the call entirely fixes the issue with minimal
overhead.

Signed-off-by: Chris N. Johnson <chris.n.johnson@intel.com>