Make reading the BSR a required part of DCPMM initialization
In situations where DCPMM media is not working / disabled, we don't
currently downgrade from DDRT large payload to DDRT small payload
as expected. This information (whether media is disabled) is only
found in the boot status register (BSR), which we currently don't read
in unless a command specifically asks for it. This change reads the
BSR early in DCPMM initialization so all subsequent calls can benefit
from the information.

This change exposed a minor bug in our PBR implementation
(-v/-smbus/-ddrt flags are recorded but not replayed correctly) that
incorrectly caused our internal smoke tests to fail. This commit
temporarily disables the offending test case and it will be re-enabled
in a future release along with fixing PBR.

A related change is in DeterminePassThruMethod(). We now correctly return
errors when the internal logic fails to determine a method to interface
with the DCPMM. The associated references are updated accordingly.

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>