Ensure cleanup functions in NvmTables file free the actual pointer
Some cleanup functions in NvmTables file such as FreeParsedPcat, FreeParsedPmtt
and FreeParsedNfit take in a pointer to a structure as argument to free the
contents of the structure. These functions call free on the pointer passed in
when the caller still has the pointer. This causes segmentation fault when the
caller calls these functions again. This was seen when there is a test case which
corrupts serial number on dimms. The fix implemented here is to pass in the address
of pointer to these functions

Signed-off-by: James Sushanth Anandraj <james.sushanth.anandraj@intel.com>