Update return code in initialize namespace function
Initialize namespace function detects if Label Storage
area is not initialized for dimms. This also sets a
return code of EFI_NOT_FOUND in case any of the
label storage area is not initialized.

In one of our test cases, there are namespaces initialized
on DCPMMs and the LSA data is then deleted for one DCPMM.
Currently we are not able to run show namespace because
the return code from ReadLabelStorageArea() is not success
for this case of an uninitialized LSA. This should not be a
failure condition as we are still able to communicate
with the DCPMM. This change removes the premature setting
of this failing return code.

Signed-off-by: James Sushanth Anandraj <james.sushanth.anandraj@intel.com>