Fix for SKU limit calculation when moving from 2LM mode to 2LM with goal request for a subset of DIMMs.
Earlier specifying a subset of DIMMs was possible only when all the unconfigured DIMMs
were targeted. But, with the latest support added for '-dimm' option even subset of DIMMs
which are already in configured state can be specified under specific conditions. This fix
makes sure that in such cases, SKU limit calculations will exclude DIMMs which are not part
of the create goal request instead of using 'Configured' state as a filter.

Signed-off-by: Sandesh Shetty <sandesh.shetty@intel.com>