Treat BTT map entries in the initial state as valid entries
To align with the UEFI spec, this change updates our BTT namespace
algorithm to treat map entries that are in the initial state
(error == 0 and zero == 0) as valid entries. Normally the LBA provided
is used as an offset into the map table, which then would provide a new
offset into the namespace to look up data with. Previously we treated
map entries in the initial state the same as the zero state, no
initialized data to return (all zeros). Now, if the map entry is in the
initial state, we are to use the pre-map LBA provided as an
offset to the namespace directly and return the data at that location.

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>