Deprecate -recover flag from FW update command, clean up code
The CLI -recover flag is now no longer required to update
"non-functional" DCPMMs, however we are maintaining the behavior
for backwards compatibility. The filtering logic for this behavior
is removed from LoadCommand.c in this change and will be
better implemented in VerifyTargetDimms() shortly.

Due to a number of defects resulting from inconsistent behavior
in the firmware update routines, this change merges and simplifies
the firmware update code and places it alongside the other firmware
commands in Dimm.c.

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>