Convert VerifyTargetDimms parameters to bitfield
This change introduces the use of a bitfield to indicate what
DCPMMs should be filtered out from the list of DCPMMs that
VerifyTargetDimms ultimately provides to a caller. It should be
much easier now to read what is getting sent to VerifyTargetDimms
and add additional parameters for filtering.

Also change firmware update behavior to include non-functional
DCPMMs in normal operation and maintain backwards compatible
behavior of "-recover" flag (only run on non-functional DCPMMs).

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>