Fix for platform hang during boot in cases where a DIMM is part of more than 2 regions.
We have 2 data structures which hold the list of interleave sets:
One is a global list and the other is maintained per DIMM related
regions. This fix is for the DIMM specific array of size 2
(MAX_IS_PER_DIMM) to hold only the healthy or mapped regions which
can be maximum 2 intereleave sets per DIMM. While the other global
list will hold all interleave sets including the broken ones too.

Signed-off-by: Sandesh Shetty <sandesh.shetty@intel.com>