Add new CCUR and COUT status codes for partial mapping of DCPMMs due to a population issue
The CCUR Configuration Status code 15 and COUT Validation Status code 8 were
added to specify the case when the DCPMM persistent memory is mapped, but
the volatile memory is not mapped due to a population issue.  These status
codes occur when a valid ADx1+MM configuration was applied, and one of the
DCPMMs goes bad and causes other DCPMMs to be in population violation.  In
this case, the ADx1 portion of the DCPMMs in population violation are still
accessible, but the volatile memory is not.

The show-device field PopulationViolationState was also renamed in the
output to PopulationViolation and updated to output 'Yes' or 'No' to align
with the HII output and maintain consistency.

Signed-off-by: Chris N. Johnson <chris.n.johnson@intel.com>
Signed-off-by: Sandesh Shetty <sandesh.shetty@intel.com>