Use an accurate method for determining DCPMM interface availability
Validation testing exposed that the BSR bit that we were using to
determine whether DDRT was accessible or not (PCR_UNLOCKED) could be set even
if DDRT was available.

Therefore, the only reliable method to do this is to send a basic command over
the interfaces we care about. At this point, the only way to do this is to
override the CLI-specified "-ddrt"/"-smbus" flags. This change creates a more
cluttered DCPMM initialization routine, but which gives the desired outcome.
A separate feature request proposes some additional ways of doing this
determination, which we will investigate at a later date.

In further testing these changes, we discovered that DeterminePassThruMethod()
needs an additional <valid> state of only "-ddrt" passed in via the CLI.

Additionally, this change removes various unused references to
EFI_DCPMM_CONFIG_TRANSPORT_ATTRIBS for better future readability.

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>