Change cmake to generate output list, had None. so sections not empty
The a2x executable rejects empty sections so the string "None." was
added.

In the CMakeLists.txt file, two lists where used previously which
needed to be updated in parallel or else manpages could be generated
incorrectly. Now one listed is created from the other so only one
list is needed in the file.

Signed-off-by: Steven Pontsler <steven.pontsler@intel.com>