Allow windows user to specify the same file name for dump system command multiple times
In os_efi_simple_file_protocol layer when EFI_FILE_MODE_CREATE is passed
and _O_CREAT flag is not applied for file_opened with proper pmode
we see that dump command fails with a file error when running the command
a second time with the same file_name

This flag is passed for linux in the same case and the dump file is overwritten
instead of returning error. With this change this flag is passed and the dump
file is overwritten instead of returning error if created by the same command.
If the file name passed is read only then we would still return error.

Signed-off-by: James Sushanth Anandraj <james.sushanth.anandraj@intel.com>