Clean up status reporting for firmware update command
Another issue was the object status for a particular PMem module
was getting overwritten by the higher level FwUpdate() function.
Changed the code to just return a proper NvmStatus and let the
caller decide what to do with it.

Moved MatchCliReturnCode() to a common code location so every function
in our codebase can use it, not just the CLI. Keep existing naming
for now.

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>