Enabled secure erase and changing master passphrase on specified PMem module(s)
The fix is to allow secure erase and changing master passphrase if master
passphrase is enabled on a subset of PMem modules. Removed the unwanted
check for population violation in SetDimm command. Also, fixed static analysis
bugs in both SetDimm and DeleteDimmm commands where PMem module list was
being populated for the second time without freeing it.

Signed-off-by: Sandesh Shetty <sandesh.shetty@intel.com>