Ignore media disabled devices when enumerating regions and namespaces
When we have a mix of dimms with some being media disabled due to fatal media error,
we can show regions and namespaces on dimms which are not media disabled by ignoring
media disabled dimms when we read pcd data or lsa data.

We still would not be able to display health state of the region or namespace on media
disabled dimm because we can't communicate with it.

Signed-off-by: James Sushanth Anandraj <james.sushanth.anandraj@intel.com>