Do not allow ARS issues to block fw update
There are a few different error scenarios that can occur related with ARS and
firmware update. This changeset attempts to explain and rectify them with a
simple change: logging but ignoring them. Additionally, this change extends the
number of retries allowed to account for repeated restarts of ARS during the
course of the many-packet small payload firmware update flow.

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>