Update to use PCAT revision for PCD and CIN headers when applying goals.
Previously, PCD CCUR revision was used when creating goals. And PCAT
revision was used only when PCD CCUR table is missing. With these
changes only PCAT revision will be used always. This is to make sure
provisioning is not impacted when the user downgrades to a lower BIOS
version with a different interface revision supported.

Signed-off-by: Sandesh Shetty <sandesh.shetty@intel.com>