Fix to calculate NM:FM ratio correctly when a subset of PMem modules are targeted for 100%MM.
This is to make sure NM:FM ratio is calculated correctly when only a
subset of PMem modules are targeted for 100% Memory Mode. Targeting a
subset of PMem modules for 100% MemoryMode is allowed if all unspecified
PMem modules are configured for 100% MemoryMode only. Volatile partitions
on these unspecified PMem modules needs to be accounted for in this case.

Signed-off-by: Sandesh Shetty <sandesh.shetty@intel.com>