Adjust show pcd command to skip media disabled modules
Instead of causing a segfault when executed on Linux, we should skip
over the media disabled module and print out a helpful explanatory
message when we finish CLI execution.

This quick check and error message creation should be incorporated
into VerifyTargetDimms.

Added a flag to the printer logic to not print a success message
if desired. Some commands want this (show -pcd here), but other
commands ("set") that don't produce output normally would benefit
from having a success message.

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>