Reduce capacity with SKU limited CPU when subset of PMem modules targeted
Targeting a subset of PMem modules for 100% AppDirect Non-Interleaved mode is
allowed if all the unspecified PMem modules already have 100% AppDirect
Non-Interleaved regions. With a SKU limited CPU, the capacity of the targeted
PMem modules should be reduced if the total AppDirect capacity requested on the
targeted ones plus the total AppDirect capacity already mapped on the unspecified
ones exceed the CPU SKU limit.

Signed-off-by: Sandesh Shetty <sandesh.shetty@intel.com>