Enhance Clear FatalMediaError message.
After injecting and then clearing a FatalMediaError, the show memory
resources command cannot retrieve memory resources info. This is
correct behavior, but lack of information from successful clearing of
the error leads to confusion. This change produces an explicit warning
to user that reboot is required for platform changes to take effect.

Signed-off-by: MacKevin Fey <mackevin.fey@intel.com>