Provide better BIOS table incompatibility reason to end user
Expose Parse****Table() failures to upper level functions (have them
return an EFI_STATUS).

Create a new function for providing a basic mapping between EFI_STATUS
codes and a reasonable guess as to the root cause
(GuessNvmStatusFromReturnCode()). This is definitely not a good or
reliable guess, but since some of our API functions don't have more
verbose pCommandStatus parameters, we only have EFI_STATUS codes
to work with.

CLI initialization now checks the driver return code and prints
out an appropriate error, for all CLI calls. However, this is
implemented for OS only, as the UEFI driver will always match
the UEFI platform firmware version in end-customer scenarios as
it is stitched in.

Nothing is intended to be changed for parsing the PCD, since we
still want to allow clearing the PCD in case of corruption and
if there is a version mismatch, the BIOS tables will reflect it
first (and better).

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>