Use proper link notation in documentation
Searched for all relevant section titles in our documentation
and made sure they used proper link notation for AsciiDoctor.
The PDF output looks as expected with light blue text and a working
HTML-style link that goes immediately to the referenced section.

Confirmed that the underline issues in the manpages are fixed
and that the text used doesn't look bad. ("<<Foo>>" in asciidoctor
looks like "[Foo]" in the manpage. Not great, but it'll work)

Additionally removed references to the former Set Power
Management Policy command.

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>