Call driver stop when driver start fails
When our DriverBindingStart() call fails, UEFI does not call our
DriverBindingStop() function, presumably because it was never started!
This change calls DriverBindingStop() if DriverBindingStart() fails
for any reason.

There might be cases where this naive implementation doesn't work,
but I can't think of them currently. As long as you ignore errors
and BindingStop() in the normal scenario cleans up after everything
that was initialized, then I don't see why it wouldn't continue to
do the same when we only partially initialized if it ignores any
errors when uninitializing (which it does).

It looks like we already do this for DriverLoad() and DriverUnload(),
so there's no need to update those functions.

Also clean up CHECK_RESULT_CONTINUE macro to not overwrite
the existing return code!

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>