Clean up pCommandStatus memory management by callers
  * Remove redundant AllocateZeroPool calls (they are done in
      InitializeCommandStatus)
  * Can't use FREE_POOL_SAFE since there's a possibility of leaking
     memory

Investigate and fix the root cause of the assert, which is
that CatSPrintNCopy overruns the end of the buffer if DestString has
garbage in it (which makes StringSize very long). I removed all of
that since it is not necesssary and replaced it with StrCatS, which
understandably still fails if garbage is provided (it's hard to
concatenate to a string that exceeds the maximum length of a buffer!),
but at least it doesn't assert! Update the documentation for
CatSPrintNCopy as well so that it's accurate.

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>