Skip enabled check and allow default on FIS 3.2
Starting with FIS 3.2, the flag indicating that the master passphrase
is enabled is not changed until the master passphrase has been changed
from the default. Thus ipmctl needs to not check the flag before
attempting to set the master passphrase when the firmware supports
FIS 3.2 or higher.

Remove check for default option when not setting master passphrase
as it is not useful. No change here for FIS 3.2.

Signed-off-by: Steven Pontsler <steven.pontsler@intel.com>