Check media disabled dimms during load config from file
In load config from file function there was call to create
goal config for every socket. But in the create goal config
there is a check for media disabled dimms on all sockets so
if there is a disabled dimm on any socket the call for create
goal config on first socket fails. So to discover the correct
source of failure a check for media disabled dimms is added
at the beginning of load config from file just like create goal.

Signed-off-by: James Sushanth Anandraj <james.sushanth.anandraj@intel.com>