Update size calculations for goal data loaded from file
When converting the volatile memory size for goal data loaded
from file a roundup function was used to round up to nearest percent.
This resulted in reduced app direct region size. Changing the roundup
to round closest fixed this issue.

Given that when dumping a goal the sizes are always dumped in GiB and when
read back, percentage values are calculated in Integer. We should expect ~0.5%
difference in sizes between regions created by command line and from dumping
and reloading the goal.

Signed-off-by: James Sushanth Anandraj <james.sushanth.anandraj@intel.com>