Show master passphrase set option if secure erase is enabled
If the secure erase feature is available, the user needs to be able
to set the master passphrase as well.

On FIS 3.2 or later, the master passphrase needs to be changed from the
default passphrase before secure erase can be used. The bit that we were using
previously to indicate whether master passphrase could be modified,
MasterPassphraseEnabled, is now not set until the default passphrase is
changed. So we need to include whether secure erase is available in our
calculation of whether to allow modification of the master passphrase.

Signed-off-by: Steven Pontsler <steven.pontsler@intel.com>