Print usage using ESXi error format when esx output is requested
ESXCLI integration expects output in a certain format and prints an
error message if it receives output that is not in that format. When
ipmctl prints usage it is not the correct format. Now when esx
output is requested via the -output option the usage will be printed
in the esx error format so ESXCLI will print the message from ipmctl.

Signed-off-by: Steven Pontsler <steven.pontsler@intel.com>