Allow PMTT table to be missing
Some BIOS vendors did not implement a PMTT table for the Purley
platform. In order to maintain backwards compatibility, this change
attempts to allow for the PMTT table to be missing (and show
warnings / "unknown" as appropriate) without changing the overall
functionality that we've added since the earlier releases of ipmctl.

Specifically, in show -memoryresources this change will show "Unknown"
for all DDR4 capacities and the resulting computed total capacities,
since they depend on an unknown DDR4 value. For create -goal, DDR4
capacity is used to make sure we don't exceed the socket sku limit
during provisioning, but for this change we will just show a warning
that BIOS could reject the request upon reboot.

For non-Purley platforms, we made the PMTT table a requirement.
In driver init, we now error out if we are missing a PMTT table *and* are on a
non-Purley platform. Because of this, this change removes many unneeded
references to GetAcpiPMTT because they were only checking the PMTT table
for NULL or 0.1 versions and skipping execution. Now, just checking
the parsed PMTT table (only 0.2 is parsed (non-Purley)) for NULL
is sufficient and equivalent.

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>