Revert ipmctl 3.0 long op status behavior back to ipmctl 2.0
Due to various suggestions by customers and representatives, we made
some changes in the past year to clarify what various long op status
codes meant. Unfortunately, this change in behavior is not universally
accepted, and we've learned it is usually better to keep things
consistent from generation to generation.

Of note is the decision to use our old method to determine ARSStatus
state from the Get Address Range Scrub command instead of Get Long Operation
Status command. This is a little more work, but is consistent with ipmctl 2.0.

However, this change continues forward in consolidating our code, and many
redundant string and function definitions are moved into common locations.

This change will affect our PAL recordings, but there are other
changes to the PAL system configuration that unfortunately need to
be included in this submission too because of the way PAL/PBR are designed.

Not sure why OverwriteStatus is "Not started" on some of the PMem
modules on both the Linux and Windows machines. This is from long
op status not getting set at all, but I'm not sure why that is.
There is an AppDirectNotInterleaved region on each PMem module (which
should trigger ARS) and I've restarted the systems, but to no avail.

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>