Use SMBIOS table for topology if PMTT is invalid or missing
Some BIOS versions do not produce a PMTT table or do not properly set
the handle field. This change works around needing the PMTT table for
the most part for show -topology. Not all fields are available such as
SocketId for regular DDR. This means if the -socket is used that DDR
dimms will filtered out.

Signed-off-by: Steven Pontsler <steven.pontsler@intel.com>