Allow start diagnostic to run on all PMem modules
A change was made in build 274 that correctly made unmanageable
PMem modules not get added to the (functional) Dimms list. This
caused any CLI commands that only pulled modules from this functional
list to reject specified PMem modules that were not on the list.
In some cases this is ok and isn't worth changing the code over,
but for others (start -diagnostic), we want to be able to run
on these unmanageable PMem modules, so this change enables the CLI
command to allow all PMem modules to be specified.

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>