Fix numerous memory leaks
FreeLsaSafe was confusing the LSA indices (called Index) with the
incrementation variable, also called Index. This change
updates the incrementation variable to IndexIdx to correct the
memory leak and to hopefully make it a little clearer.

While freeing allocated NvmTables, there were some mislabeled fields.

NvmDimmDriver.c
  * Index for writing to device path wasn't incrementing. I think
    our driver was still able to function correctly, but the pointers
    used were overwritten and lost.
  * Expanded the scope of init code that runs during OS so we can make
    code more common and be sure to free memory on exit like UEFI does.

Needed to adjust Set PCD as it was reading over the end of the allocated buffer.

HealthStateReason is cleaned up, and I verified it is working properly for
healthy, media disabled, and multiple health reasons at the same time.

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>