Allow create goal to handle media disabled modules correctly
Make ApplyGoalConfigsToDimms compatible with VerifyTargetDimms and have it
use the specified subset of PMem modules.

Exclude media not accessible modules when we write out the PCD. The
existing code is not good practice, but rewriting the APIs would take
some time as well.

Update a few commands to only use media accessible PMem modules. I confirmed
that the rest of the PCD-relevant commands (regions, namespaces, and PCD) work
as expected.

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>