Make ipmctl more resilient to unhealthy PMem modules
Only zero out headers for PCD and LSA when calling delete -pcd. Will reduce
runtime on non-functional PMem modules from 10 minutes down to 10 seconds. In
FwCmdSetPlatformConfigData, use the same flow for PCD and LSA when we
set PcdSize, there is no reason for there to be a difference.

For non-functional PMem modules, allow delete -goal to work on them and
exclude them from checking for existing goals on the socket (since we are
not going to provision them with create -goal).

Don't bail out when there are errors retrieving PCD during Retrieve Goal
Configs. However, make sure to preserve the behavior used in the past for
when EFI_NO_RESPONSE is returned ("One or more dimms are busy"...)

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>