Correct multiple NVM API behavior deficiencies
By default the nvm_get_region* calls of the NVM API would use NFIT
instead of PCD. This causes problems when new namespaces are created
because the FreeCapacity in that region would not get updated. This
change updates the default behavior for the standard nvm_get_region*
API calls to use the PCD instead of the NFIT, and brings it into
alignment with how the CLI has behaved for the last 3 years. We see
no degradation of behavior from making this change, other than
if the PCD is corrupted, we won't get any region information, which is
an unimportant scenario.

In addition, this change enables the NVM API to set the default
passthrough interfaces to be the same as the OS CLI. Previously, in OS
a call to the nvm_get_region* commands would take 10s of seconds to
initialize each PMem module because it was using the Large Payload
interface.

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>