Correct link removal behavior for preferences file
We allow a user to have a preferences file that is "local", in the same
directory as ipmctl execution. If ipmctl is executed from an accessible
location, then there's a chance a malicious actor could install a link
with the same name as the ipmctl preferences that points to an important
admin-protected file in some other location. We checked in a change a few
months ago to protect against this attack vector by deleting any shortcut
file with the same name as the preference file in the local directory, but
it deleted normal files as well. This change corrects this behavior by
using the proper APIs to check if a file is a shortcut.

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>